+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @aparte-workspace/docs
 
+## 0.0.27
+
+### Patch Changes
+
+- Updated dependencies [6d3272e]
+  - @aparte/core@0.16.11
+  - @aparte/locale-fr@0.16.11
+  - @aparte/plugin-shiki@0.16.11
+
+## 0.0.26
+
+### Patch Changes
+
+- Updated dependencies [33b8cc0]
+- Updated dependencies [36af623]
+  - @aparte/core@0.16.10
+  - @aparte/locale-fr@0.16.10
+  - @aparte/plugin-shiki@0.16.10
+
 ## 0.0.25
 
 ### Patch Changes