+++ apps/examples/vanilla-dist/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @aparte-workspace/demo-vanilla
 
+## 0.0.27
+
+### Patch Changes
+
+- Updated dependencies [6d3272e]
+  - @aparte/core@0.16.11
+
+## 0.0.26
+
+### Patch Changes
+
+- Updated dependencies [33b8cc0]
+- Updated dependencies [36af623]
+  - @aparte/core@0.16.10
+
 ## 0.0.25
 
 ### Patch Changes