+++ packages/plugins/streaming-markdown/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @aparte/plugin-streaming-markdown
 
+## 0.16.11
+
+## 0.16.10
+
 ## 0.16.9
 
 ## 0.16.8