+++ packages/providers/ai/transformers/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @aparte/provider-transformers
 
+## 0.16.11
+
+## 0.16.10
+
 ## 0.16.9
 
 ### Patch Changes