+++ packages/wrappers/svelte/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @aparte/svelte
 
+## 0.16.11
+
+## 0.16.10
+
 ## 0.16.9
 
 ### Patch Changes